Allow Envoy mobile to build sans YAML. e2e tests will not pass until https://github.com/envoyproxy/envoy/pull/26451 lands
