+++ .github/workflows/mobile-compile_time_options.yml
@@ -43,6 +43,21 @@ jobs:
             --@envoy//bazel:http3=False \
             --@com_envoyproxy_protoc_gen_validate//bazel:template-flavor= \
             //test/cc/...
+    - name: 'Build Java sans yaml'
+      if: steps.should_run.outputs.run_ci_job == 'true'
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      # Envoy Mobile build verifies that the build configuration where YAML is disabled.
+      run: |
+        # TODO(alyssawilk) run relevant tests once runtime changes land
+        cd mobile && ./bazelw build \
+            $([ -z $GITHUB_TOKEN ] || echo "--config=remote-ci-linux") \
+            --define=envoy_mobile_request_compression=disabled \
+            --define=envoy_enable_http_datagrams=disabled \
+            --define=google_grpc=disabled \
+            --define=envoy_yaml=disabled \
+            --@com_envoyproxy_protoc_gen_validate//bazel:template-flavor= \
+            //test/java/...
   swift_build:
     if: github.repository == 'envoyproxy/envoy'
     name: swift_build
