+++ bazel/envoy_select.bzl
@@ -87,6 +87,13 @@ def envoy_select_enable_http3(xs, repository = ""):
         "//conditions:default": xs,
     })
 
+# Selects the given values if yaml is enabled in the current build.
+def envoy_select_enable_yaml(xs, repository = ""):
+    return select({
+        repository + "//bazel:disable_yaml": [],
+        "//conditions:default": xs,
+    })
+
 # Selects the given values if HTTP datagram support is enabled in the current build.
 def envoy_select_enable_http_datagrams(xs, repository = ""):
     return select({
