+++ source/common/formatter/substitution_formatter.cc
@@ -150,8 +150,13 @@ std::string JsonFormatterImpl::format(const Http::RequestHeaderMap& request_head
   const ProtobufWkt::Struct output_struct = struct_formatter_.format(
       request_headers, response_headers, response_trailers, stream_info, local_reply_body);
 
+#ifdef ENVOY_ENABLE_YAML
   const std::string log_line =
       MessageUtil::getJsonStringFromMessageOrError(output_struct, false, true);
+#else
+  IS_ENVOY_BUG("Json support compiled out");
+  const std::string log_line = "";
+#endif
   return absl::StrCat(log_line, "\n");
 }
 
@@ -1897,12 +1902,16 @@ MetadataFormatter::formatMetadata(const envoy::config::core::v3::Metadata& metad
   if (value.kind_case() == ProtobufWkt::Value::kStringValue) {
     str = value.string_value();
   } else {
+#ifdef ENVOY_ENABLE_YAML
     ProtobufUtil::StatusOr<std::string> json_or_error =
         MessageUtil::getJsonStringFromMessage(value, false, true);
     ENVOY_BUG(json_or_error.ok(), "Failed to parse json");
     if (json_or_error.ok()) {
       str = json_or_error.value();
     }
+#else
+    IS_ENVOY_BUG("Json support compiled out");
+#endif
   }
   truncate(str, max_length_);
   return str;
@@ -2097,10 +2106,12 @@ ProtobufWkt::Value FilterStateFormatter::formatValue(const Http::RequestHeaderMa
     return unspecifiedValue();
   }
 
+#ifdef ENVOY_ENABLE_YAML
   ProtobufWkt::Value val;
   if (MessageUtil::jsonConvertValue(*proto, val)) {
     return val;
   }
+#endif
   return unspecifiedValue();
 }
