+++ source/common/protobuf/utility.cc
@@ -20,7 +20,6 @@
 #include "udpa/annotations/status.pb.h"
 #include "validate/validate.h"
 #include "xds/annotations/v3/status.pb.h"
-#include "yaml-cpp/yaml.h"
 
 using namespace std::chrono_literals;
 
@@ -33,92 +32,6 @@ absl::string_view filenameFromPath(absl::string_view full_path) {
     return full_path;
   }
   return full_path.substr(index + 1, full_path.size());
-}
-
-void blockFormat(YAML::Node node) {
-  node.SetStyle(YAML::EmitterStyle::Block);
-
-  if (node.Type() == YAML::NodeType::Sequence) {
-    for (const auto& it : node) {
-      blockFormat(it);
-    }
-  }
-  if (node.Type() == YAML::NodeType::Map) {
-    for (const auto& it : node) {
-      blockFormat(it.second);
-    }
-  }
-}
-
-ProtobufWkt::Value parseYamlNode(const YAML::Node& node) {
-  ProtobufWkt::Value value;
-  switch (node.Type()) {
-  case YAML::NodeType::Null:
-    value.set_null_value(ProtobufWkt::NULL_VALUE);
-    break;
-  case YAML::NodeType::Scalar: {
-    if (node.Tag() == "!") {
-      value.set_string_value(node.as<std::string>());
-      break;
-    }
-    bool bool_value;
-    if (YAML::convert<bool>::decode(node, bool_value)) {
-      value.set_bool_value(bool_value);
-      break;
-    }
-    int64_t int_value;
-    if (YAML::convert<int64_t>::decode(node, int_value)) {
-      if (std::numeric_limits<int32_t>::min() <= int_value &&
-          std::numeric_limits<int32_t>::max() >= int_value) {
-        // We could convert all integer values to string but it will break some stuff relying on
-        // ProtobufWkt::Struct itself, only convert small numbers into number_value here.
-        value.set_number_value(int_value);
-      } else {
-        // Proto3 JSON mapping allows use string for integer, this still has to be converted from
-        // int_value to support hexadecimal and octal literals.
-        value.set_string_value(std::to_string(int_value));
-      }
-      break;
-    }
-    // Fall back on string, including float/double case. When protobuf parse the JSON into a message
-    // it will convert based on the type in the message definition.
-    value.set_string_value(node.as<std::string>());
-    break;
-  }
-  case YAML::NodeType::Sequence: {
-    auto& list_values = *value.mutable_list_value()->mutable_values();
-    for (const auto& it : node) {
-      *list_values.Add() = parseYamlNode(it);
-    }
-    break;
-  }
-  case YAML::NodeType::Map: {
-    auto& struct_fields = *value.mutable_struct_value()->mutable_fields();
-    for (const auto& it : node) {
-      if (it.first.Tag() != "!ignore") {
-        struct_fields[it.first.as<std::string>()] = parseYamlNode(it.second);
-      }
-    }
-    break;
-  }
-  case YAML::NodeType::Undefined:
-    throw EnvoyException("Undefined YAML value");
-  }
-  return value;
-}
-
-void jsonConvertInternal(const Protobuf::Message& source,
-                         ProtobufMessage::ValidationVisitor& validation_visitor,
-                         Protobuf::Message& dest) {
-  Protobuf::util::JsonPrintOptions json_options;
-  json_options.preserve_proto_field_names = true;
-  std::string json;
-  const auto status = Protobuf::util::MessageToJsonString(source, &json, json_options);
-  if (!status.ok()) {
-    throw EnvoyException(fmt::format("Unable to convert protobuf message to JSON string: {} {}",
-                                     status.ToString(), source.DebugString()));
-  }
-  MessageUtil::loadFromJson(json, dest, validation_visitor);
 }
 
 // Logs a warning for use of a deprecated field or runtime-overridden use of an
@@ -236,101 +149,6 @@ size_t MessageUtil::hash(const Protobuf::Message& message) {
   }
 
   return HashUtil::xxHash64(text_format);
-}
-
-void MessageUtil::loadFromJson(const std::string& json, Protobuf::Message& message,
-                               ProtobufMessage::ValidationVisitor& validation_visitor) {
-  bool has_unknown_field;
-  auto status = loadFromJsonNoThrow(json, message, has_unknown_field);
-  if (status.ok()) {
-    return;
-  }
-  if (has_unknown_field) {
-    // If the parsing failure is caused by the unknown fields.
-    validation_visitor.onUnknownField("type " + message.GetTypeName() + " reason " +
-                                      status.ToString());
-  } else {
-    // If the error has nothing to do with unknown field.
-    throw EnvoyException("Unable to parse JSON as proto (" + status.ToString() + "): " + json);
-  }
-}
-
-Protobuf::util::Status MessageUtil::loadFromJsonNoThrow(const std::string& json,
-                                                        Protobuf::Message& message,
-                                                        bool& has_unknown_fileld) {
-  has_unknown_fileld = false;
-  Protobuf::util::JsonParseOptions options;
-  options.case_insensitive_enum_parsing = true;
-  // Let's first try and get a clean parse when checking for unknown fields;
-  // this should be the common case.
-  options.ignore_unknown_fields = false;
-  // Clear existing values (if any) from the destination message before serialization.
-  message.Clear();
-  const auto strict_status = Protobuf::util::JsonStringToMessage(json, &message, options);
-  if (strict_status.ok()) {
-    // Success, no need to do any extra work.
-    return strict_status;
-  }
-  // If we fail, we see if we get a clean parse when allowing unknown fields.
-  // This is essentially a workaround
-  // for https://github.com/protocolbuffers/protobuf/issues/5967.
-  // TODO(htuch): clean this up when protobuf supports JSON/YAML unknown field
-  // detection directly.
-  options.ignore_unknown_fields = true;
-  const auto relaxed_status = Protobuf::util::JsonStringToMessage(json, &message, options);
-  // If we still fail with relaxed unknown field checking, the error has nothing
-  // to do with unknown fields.
-  if (relaxed_status.ok()) {
-    has_unknown_fileld = true;
-    return strict_status;
-  }
-  return relaxed_status;
-}
-
-void MessageUtil::loadFromJson(const std::string& json, ProtobufWkt::Struct& message) {
-  // No need to validate if converting to a Struct, since there are no unknown
-  // fields possible.
-  loadFromJson(json, message, ProtobufMessage::getNullValidationVisitor());
-}
-
-void MessageUtil::loadFromYaml(const std::string& yaml, Protobuf::Message& message,
-                               ProtobufMessage::ValidationVisitor& validation_visitor) {
-  ProtobufWkt::Value value = ValueUtil::loadFromYaml(yaml);
-  if (value.kind_case() == ProtobufWkt::Value::kStructValue ||
-      value.kind_case() == ProtobufWkt::Value::kListValue) {
-    jsonConvertInternal(value, validation_visitor, message);
-    return;
-  }
-  throw EnvoyException("Unable to convert YAML as JSON: " + yaml);
-}
-
-void MessageUtil::loadFromFile(const std::string& path, Protobuf::Message& message,
-                               ProtobufMessage::ValidationVisitor& validation_visitor,
-                               Api::Api& api) {
-  const std::string contents = api.fileSystem().fileReadToEnd(path);
-  // If the filename ends with .pb, attempt to parse it as a binary proto.
-  if (absl::EndsWithIgnoreCase(path, FileExtensions::get().ProtoBinary)) {
-    // Attempt to parse the binary format.
-    if (message.ParseFromString(contents)) {
-      MessageUtil::checkForUnexpectedFields(message, validation_visitor);
-    }
-    return;
-  }
-
-  // If the filename ends with .pb_text, attempt to parse it as a text proto.
-  if (absl::EndsWithIgnoreCase(path, FileExtensions::get().ProtoText)) {
-    if (Protobuf::TextFormat::ParseFromString(contents, &message)) {
-      return;
-    }
-    throw EnvoyException("Unable to parse file \"" + path + "\" as a text protobuf (type " +
-                         message.GetTypeName() + ")");
-  }
-  if (absl::EndsWithIgnoreCase(path, FileExtensions::get().Yaml) ||
-      absl::EndsWithIgnoreCase(path, FileExtensions::get().Yml)) {
-    loadFromYaml(contents, message, validation_visitor);
-  } else {
-    loadFromJson(contents, message, validation_visitor);
-  }
 }
 
 namespace {
@@ -493,69 +311,6 @@ void MessageUtil::recursivePgvCheck(const Protobuf::Message& message) {
   ProtobufMessage::traverseMessage(visitor, message, true);
 }
 
-std::string MessageUtil::getYamlStringFromMessage(const Protobuf::Message& message,
-                                                  const bool block_print,
-                                                  const bool always_print_primitive_fields) {
-
-  auto json_or_error = getJsonStringFromMessage(message, false, always_print_primitive_fields);
-  if (!json_or_error.ok()) {
-    throw EnvoyException(json_or_error.status().ToString());
-  }
-  YAML::Node node;
-  TRY_ASSERT_MAIN_THREAD { node = YAML::Load(json_or_error.value()); }
-  END_TRY
-  catch (YAML::ParserException& e) {
-    throw EnvoyException(e.what());
-  }
-  catch (YAML::BadConversion& e) {
-    throw EnvoyException(e.what());
-  }
-  catch (std::exception& e) {
-    // Catch unknown YAML parsing exceptions.
-    throw EnvoyException(fmt::format("Unexpected YAML exception: {}", +e.what()));
-  }
-  if (block_print) {
-    blockFormat(node);
-  }
-  YAML::Emitter out;
-  out << node;
-  return out.c_str();
-}
-
-ProtobufUtil::StatusOr<std::string>
-MessageUtil::getJsonStringFromMessage(const Protobuf::Message& message, const bool pretty_print,
-                                      const bool always_print_primitive_fields) {
-  Protobuf::util::JsonPrintOptions json_options;
-  // By default, proto field names are converted to camelCase when the message is converted to JSON.
-  // Setting this option makes debugging easier because it keeps field names consistent in JSON
-  // printouts.
-  json_options.preserve_proto_field_names = true;
-  if (pretty_print) {
-    json_options.add_whitespace = true;
-  }
-  // Primitive types such as int32s and enums will not be serialized if they have the default value.
-  // This flag disables that behavior.
-  if (always_print_primitive_fields) {
-    json_options.always_print_primitive_fields = true;
-  }
-  std::string json;
-  if (auto status = Protobuf::util::MessageToJsonString(message, &json, json_options);
-      !status.ok()) {
-    return status;
-  }
-  return json;
-}
-
-std::string MessageUtil::getJsonStringFromMessageOrError(const Protobuf::Message& message,
-                                                         bool pretty_print,
-                                                         bool always_print_primitive_fields) {
-  auto json_or_error =
-      getJsonStringFromMessage(message, pretty_print, always_print_primitive_fields);
-  return json_or_error.ok() ? std::move(json_or_error).value()
-                            : fmt::format("Failed to convert protobuf message to JSON string: {}",
-                                          json_or_error.status().ToString());
-}
-
 void MessageUtil::unpackTo(const ProtobufWkt::Any& any_message, Protobuf::Message& message) {
   if (!any_message.UnpackTo(&message)) {
     throw EnvoyException(fmt::format("Unable to unpack as {}: {}",
@@ -573,35 +328,6 @@ absl::Status MessageUtil::unpackToNoThrow(const ProtobufWkt::Any& any_message,
   }
   // Ok Status is returned if `UnpackTo` succeeded.
   return absl::OkStatus();
-}
-
-void MessageUtil::jsonConvert(const Protobuf::Message& source, ProtobufWkt::Struct& dest) {
-  // Any proto3 message can be transformed to Struct, so there is no need to check for unknown
-  // fields. There is one catch; Duration/Timestamp etc. which have non-object canonical JSON
-  // representations don't work.
-  jsonConvertInternal(source, ProtobufMessage::getNullValidationVisitor(), dest);
-}
-
-void MessageUtil::jsonConvert(const ProtobufWkt::Struct& source,
-                              ProtobufMessage::ValidationVisitor& validation_visitor,
-                              Protobuf::Message& dest) {
-  jsonConvertInternal(source, validation_visitor, dest);
-}
-
-bool MessageUtil::jsonConvertValue(const Protobuf::Message& source, ProtobufWkt::Value& dest) {
-  Protobuf::util::JsonPrintOptions json_options;
-  json_options.preserve_proto_field_names = true;
-  std::string json;
-  auto status = Protobuf::util::MessageToJsonString(source, &json, json_options);
-  if (!status.ok()) {
-    return false;
-  }
-  bool has_unknow_field;
-  status = MessageUtil::loadFromJsonNoThrow(json, dest, has_unknow_field);
-  if (status.ok() || has_unknow_field) {
-    return true;
-  }
-  return false;
 }
 
 ProtobufWkt::Struct MessageUtil::keyValueStruct(const std::string& key, const std::string& value) {
@@ -709,15 +435,31 @@ bool redactTypedStruct(Protobuf::Message* message, const char* typed_struct_type
       message, ancestor_is_sensitive, typed_struct_type,
       [message](Protobuf::Message* typed_message, const Protobuf::Reflection* reflection,
                 const Protobuf::FieldDescriptor* field_descriptor) {
+#ifdef ENVOY_ENABLE_YAML
         // To unpack a `TypedStruct`, convert the struct from JSON.
-        jsonConvertInternal(reflection->GetMessage(*message, field_descriptor),
-                            ProtobufMessage::getNullValidationVisitor(), *typed_message);
+        MessageUtil::jsonConvert(reflection->GetMessage(*message, field_descriptor),
+                                 *typed_message);
+#else
+        UNREFERENCED_PARAMETER(message);
+        UNREFERENCED_PARAMETER(typed_message);
+        UNREFERENCED_PARAMETER(reflection);
+        UNREFERENCED_PARAMETER(field_descriptor);
+        IS_ENVOY_BUG("redaction requested with JSON/YAML support removed");
+#endif
       },
       [message](Protobuf::Message* typed_message, const Protobuf::Reflection* reflection,
                 const Protobuf::FieldDescriptor* field_descriptor) {
-        // To repack a `TypedStruct`, convert the message back to JSON.
-        jsonConvertInternal(*typed_message, ProtobufMessage::getNullValidationVisitor(),
-                            *(reflection->MutableMessage(message, field_descriptor)));
+  // To repack a `TypedStruct`, convert the message back to JSON.
+#ifdef ENVOY_ENABLE_YAML
+        MessageUtil::jsonConvert(*typed_message,
+                                 *(reflection->MutableMessage(message, field_descriptor)));
+#else
+        UNREFERENCED_PARAMETER(message);
+        UNREFERENCED_PARAMETER(typed_message);
+        UNREFERENCED_PARAMETER(reflection);
+        UNREFERENCED_PARAMETER(field_descriptor);
+        IS_ENVOY_BUG("redaction requested with JSON/YAML support removed");
+#endif
       });
 }
 
@@ -796,53 +538,6 @@ void MessageUtil::wireCast(const Protobuf::Message& src, Protobuf::Message& dst)
   // this can fail.
   if (!dst.ParseFromString(src.SerializeAsString())) {
     throw EnvoyException("Unable to deserialize during wireCast()");
-  }
-}
-
-std::string MessageUtil::sanitizeUtf8String(absl::string_view input) {
-  if (!Runtime::runtimeFeatureEnabled(
-          "envoy.reloadable_features.service_sanitize_non_utf8_strings")) {
-    return std::string(input);
-  }
-
-  // This returns the original string if it was already valid, and returns a pointer to
-  // `result.data()` if it needed to coerce. The coerced string is always
-  // the same length as the source string.
-  //
-  // Initializing `result` to `input` ensures that `result` is correctly sized to receive the
-  // modified string, or in the case where no modification is needed it already contains the correct
-  // value, so `result` can be returned in both cases.
-  //
-  // The choice of '!' is somewhat arbitrary, but we wanted to avoid any character that has
-  // special semantic meaning in URLs or similar.
-  std::string result(input);
-  const char* sanitized = google::protobuf::internal::UTF8CoerceToStructurallyValid(
-      google::protobuf::StringPiece(input.data(), input.length()), result.data(), '!');
-  ASSERT(sanitized == result.data() || sanitized == input.data());
-
-  // Validate requirement that if the input string is returned from
-  // `UTF8CoerceToStructurallyValid`, no modification was made to result so it still contains the
-  // correct return value.
-  ASSERT(sanitized == result.data() || result == input);
-
-  return result;
-}
-
-ProtobufWkt::Value ValueUtil::loadFromYaml(const std::string& yaml) {
-  TRY_ASSERT_MAIN_THREAD { return parseYamlNode(YAML::Load(yaml)); }
-  END_TRY
-  catch (YAML::ParserException& e) {
-    throw EnvoyException(e.what());
-  }
-  catch (YAML::BadConversion& e) {
-    throw EnvoyException(e.what());
-  }
-  catch (std::exception& e) {
-    // There is a potentially wide space of exceptions thrown by the YAML parser,
-    // and enumerating them all may be difficult. Envoy doesn't work well with
-    // unhandled exceptions, so we capture them and record the exception name in
-    // the Envoy Exception text.
-    throw EnvoyException(fmt::format("Unexpected YAML exception: {}", +e.what()));
   }
 }
