+++ source/common/protobuf/utility.h
@@ -251,6 +251,7 @@ public:
    */
   static std::size_t hash(const Protobuf::Message& message);
 
+#ifdef ENVOY_ENABLE_YAML
   static void loadFromJson(const std::string& json, Protobuf::Message& message,
                            ProtobufMessage::ValidationVisitor& validation_visitor);
   /**
@@ -268,6 +269,7 @@ public:
                            ProtobufMessage::ValidationVisitor& validation_visitor);
   static void loadFromFile(const std::string& path, Protobuf::Message& message,
                            ProtobufMessage::ValidationVisitor& validation_visitor, Api::Api& api);
+#endif
 
   /**
    * Checks for use of deprecated fields in message and all sub-messages.
@@ -323,12 +325,14 @@ public:
     }
   }
 
+#ifdef ENVOY_ENABLE_YAML
   template <class MessageType>
   static void loadFromYamlAndValidate(const std::string& yaml, MessageType& message,
                                       ProtobufMessage::ValidationVisitor& validation_visitor) {
     loadFromYaml(yaml, message, validation_visitor);
     validate(message, validation_visitor);
   }
+#endif
 
   /**
    * Downcast and validate protoc-gen-validate constraints on a given protobuf.
@@ -448,6 +452,7 @@ public:
     return reflection->GetString(message, name_field);
   }
 
+#ifdef ENVOY_ENABLE_YAML
   /**
    * Convert between two protobufs via a JSON round-trip. This is used to translate arbitrary
    * messages to/from google.protobuf.Struct.
@@ -456,6 +461,7 @@ public:
    * @param source message.
    * @param dest message.
    */
+  static void jsonConvert(const Protobuf::Message& source, Protobuf::Message& dest);
   static void jsonConvert(const Protobuf::Message& source, ProtobufWkt::Struct& dest);
   static void jsonConvert(const ProtobufWkt::Struct& source,
                           ProtobufMessage::ValidationVisitor& validation_visitor,
@@ -502,6 +508,7 @@ public:
   static std::string getJsonStringFromMessageOrError(const Protobuf::Message& message,
                                                      bool pretty_print = false,
                                                      bool always_print_primitive_fields = false);
+#endif
 
   /**
    * Utility method to create a Struct containing the passed in key/value strings.
@@ -568,10 +575,12 @@ class ValueUtil {
 public:
   static std::size_t hash(const ProtobufWkt::Value& value) { return MessageUtil::hash(value); }
 
+#ifdef ENVOY_ENABLE_YAML
   /**
    * Load YAML string into ProtobufWkt::Value.
    */
   static ProtobufWkt::Value loadFromYaml(const std::string& yaml);
+#endif
 
   /**
    * Compare two ProtobufWkt::Values for equality.
