+++ source/common/upstream/health_checker_event_logger.cc
@@ -65,12 +65,14 @@ void HealthCheckEventLoggerImpl::createHealthCheckEvent(
   TimestampUtil::systemClockToTimestamp(time_source_.systemTime(), *event.mutable_timestamp());
 
   callback(event);
+#ifdef ENVOY_ENABLE_YAML
 
   // Make sure the type enums make it into the JSON
   const auto json =
       MessageUtil::getJsonStringFromMessageOrError(event, /* pretty_print */ false,
                                                    /* always_print_primitive_fields */ true);
   file_->write(fmt::format("{}\n", json));
+#endif
 }
 } // namespace Upstream
 } // namespace Envoy
