+++ source/server/server.cc
@@ -127,7 +127,8 @@ InstanceImpl::InstanceImpl(
   }
   END_TRY
   catch (const EnvoyException& e) {
-    ENVOY_LOG(critical, "error initializing configuration '{}': {}", options.configPath(),
+    ENVOY_LOG(critical, "error initializing config '{} {} {}': {}",
+              options.configProto().DebugString(), options.configYaml(), options.configPath(),
               e.what());
     terminate();
     throw;
@@ -384,13 +385,24 @@ void InstanceUtil::loadBootstrapConfig(envoy::config::bootstrap::v3::Bootstrap& 
   }
 
   if (!config_path.empty()) {
+#ifdef ENVOY_ENABLE_YAML
     MessageUtil::loadFromFile(config_path, bootstrap, validation_visitor, api);
+#else
+    if (!config_path.empty()) {
+      throw EnvoyException("Cannot load from file with YAML disabled\n");
+    }
+    UNREFERENCED_PARAMETER(api);
+#endif
   }
   if (!config_yaml.empty()) {
+#ifdef ENVOY_ENABLE_YAML
     envoy::config::bootstrap::v3::Bootstrap bootstrap_override;
     MessageUtil::loadFromYaml(config_yaml, bootstrap_override, validation_visitor);
     // TODO(snowp): The fact that we do a merge here doesn't seem to be covered under test.
     bootstrap.MergeFrom(bootstrap_override);
+#else
+    throw EnvoyException("Cannot load from YAML with YAML disabled\n");
+#endif
   }
   if (config_proto.ByteSizeLong() != 0) {
     bootstrap.MergeFrom(config_proto);
@@ -824,7 +836,9 @@ void InstanceImpl::startWorkers() {
 
 Runtime::LoaderPtr InstanceUtil::createRuntime(Instance& server,
                                                Server::Configuration::Initial& config) {
+#ifdef ENVOY_ENABLE_YAML
   ENVOY_LOG(info, "runtime: {}", MessageUtil::getYamlStringFromMessage(config.runtime()));
+#endif
   return std::make_unique<Runtime::LoaderImpl>(
       server.dispatcher(), server.threadLocal(), config.runtime(), server.localInfo(),
       server.stats(), server.api().randomGenerator(),
