+++ test/integration/BUILD
@@ -9,6 +9,7 @@ load(
     "envoy_proto_library",
     "envoy_select_admin_functionality",
     "envoy_select_enable_http3",
+    "envoy_select_enable_yaml",
     "envoy_select_envoy_mobile_listener",
     "envoy_sh_test",
 )
@@ -1000,14 +1001,12 @@ envoy_cc_test_library(
     name = "utility_lib",
     srcs = [
         "server.cc",
-        "ssl_utility.cc",
         "utility.cc",
-    ],
+    ] + envoy_select_enable_yaml(["ssl_utility.cc"]),
     hdrs = [
         "server.h",
-        "ssl_utility.h",
         "utility.h",
-    ],
+    ] + envoy_select_enable_yaml(["ssl_utility.h"]),
     data = ["//test/common/runtime:filesystem_test_data"],
     deps = [
         ":server_stats_interface",
@@ -1041,8 +1040,6 @@ envoy_cc_test_library(
         "//source/server:options_lib",
         "//source/server:process_context_lib",
         "//source/server:server_lib",
-        "//test/common/upstream:utility_lib",
-        "//test/config:utility_lib",
         "//test/mocks:common_lib",
         "//test/mocks/event:event_mocks",
         "//test/mocks/runtime:runtime_mocks",
@@ -1054,13 +1051,15 @@ envoy_cc_test_library(
         "//test/test_common:simulated_time_system_lib",
         "//test/test_common:test_time_lib",
         "//test/test_common:test_time_system_interface",
-        "//test/test_common:utility_lib",
         "@com_google_absl//absl/synchronization",
         "@envoy_api//envoy/config/bootstrap/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/transport_sockets/quic/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
-    ],
+    ] + envoy_select_enable_yaml([
+        "//test/common/upstream:utility_lib",
+        "//test/config:utility_lib",
+    ]),
 )
 
 envoy_cc_test_library(
